fix: more mod related issues

+++ subprocess/src/main/java/dev/ftb/app/data/mod/ModManifest.java
@@ -204,7 +204,7 @@ public class ModManifest {
         private final List<String> mirrors = new LinkedList<>();
         private @Nullable String sha1;
         private @Nullable String md5;
-        private int murmur = -1;
+        private long murmur = -1L;
         private long size;
 
         private boolean clientonly;
@@ -222,7 +222,7 @@ public class ModManifest {
         public List<String> getMirrors() { return mirrors; }
         public String getSha1() { return Objects.requireNonNull(sha1); }
         public String getMd5() { return Objects.requireNonNull(md5); }
-        public int getMurmur() { return murmur; }
+        public long getMurmur() { return murmur; }
         public long getSize() { return size; }
         public boolean isClientonly() { return clientonly; }
         public String getType() { return Objects.requireNonNull(type); }
